Use single file tar.gz sneakernet replication.
This modifies the sync logic to use hyperlog-sneakernet-replicator,
which performs hyperlog replication between a log and a single gzipped
tarball.

In terms of the sync user interface, this meant adding an additional
button ('Create File'), to differentiate between selecting an existing
.mapeodata to sync to, and creating a new one. One dialog would be
ideal, but Electron doesn't offer a nice Dialog API for this.