fix: Bundle all files in node_modules (fixes react-pdf)
React PDF has separate builds for node and the browser. Previously
Webpack was not bundling any modules from node_modules, so they were
being loaded by require() at runtime with Node module name resolution,
so the node bundle of @react-pdf was being loaded.

This change bundles all modules in node_modules in a single bundle. When
bundling Webpack follows the "browser" field in package.json, so the
correct build of react-pdf is bundled. The other effect of this is likely
faster initial load times on machines with slow disk access, since the
initial synchronous require() calls of all the many modules has a cost