fix: Fix startup race condition
The main process expects the renderer window to not finish loading until
the server is already listening. If the renderer window loads before
this then the server host and port is not available, and the
`did-finish-load` event does not fire in the `notifyReady()` function
(because it has already fired).

This is a pretty hacky solution, and we need a coordinated state machine
between front-end and back-end to track the server state. On mobile
the front-end is responsible for starting the server, but in electron
it is the other way around, so the solution needs to be slightly
different