fix: Fix coercing of field types
The user could change presets, such that the value of any particular
field is not guaranteed. We need to handle the case of a number field
being defined for a field whose current value is not a number. This
code tries to coerce a value to the type expected for the field, but if
coercion fails, then an empty field is shown.