Cache and eagerly render PDF pages, and support initial page > 1
Support starting the report preview at a page > 1, which requires
previous pages to be rendered first in order to establish which
observation should appear on the current page. Also adds caching of
observation PDFs (each observation is rendered to its own PDF during
preview) and eagerly rendering the next observation, to reduce delay.