fix: Queue PDF renders to use the same react-pdf instance
Previous code was causing React warnings of re-using the same context in
multiple renderers, because we created a new renderer instance for each
PDF preview doc (one for each observation). This changes the code to re-
use the same renderer, and update it with the doc for each page. It
should also be faster, since it's updating existing nodes rather than
re-creating the react tree.