fix: PDF reports (#456)
* chore(PDFReport): Code formatting

* feat(Reports): Fit images within a square

Allows easier calculation of pages prior to layout of images

* chore(Reports): Refactor PDF Report component

Avoid pass-through of props to components nested in components

* Allow navigation through entire report

* tweak CSS to show report paper outline clearly

* Smooth loading state transitions

* Cache and eagerly render PDF pages, and support initial page > 1

Support starting the report preview at a page > 1, which requires
previous pages to be rendered first in order to establish which
observation should appear on the current page. Also adds caching of
observation PDFs (each observation is rendered to its own PDF during
preview) and eagerly rendering the next observation, to reduce delay.

* chore: tidy up

* Add initialPageNumber option and add stories

* Add current page number

* chore: Small type fixes

* chore: Rename and add code comments

* fix: Queue PDF renders to use the same react-pdf instance

Previous code was causing React warnings of re-using the same context in
multiple renderers, because we created a new renderer instance for each
PDF preview doc (one for each observation). This changes the code to re-
use the same renderer, and update it with the doc for each page. It
should also be faster, since it's updating existing nodes rather than
re-creating the react tree.

* fix bug where loop could continue to infinity

* add missing dep to useEffect

* Update intl messages

* commit case change to git

* Add save button

* Add support for intl characters in PDF report

* fix webpack config to load fonts

* Add new design page navigator

* Format report toolbar and add x out of y observations count

* chore: fix types

* fix page navigator so that buttons are equally sized

* fix toolbar so that text is centered to page

* remove bottom border radius on page navigator

* 🎉 New report design & layout

* remove eager rendering (premature optimization, and caused problems)

* 🚀 page numbers in footer

* Fix re-render of PDF after cancelling edit of observation

* chore: Catch errors in PDF render and report them

* fix: Reset to page 1 when the user is on an invalid page

* chore: Change startPage to actually mean statePage (not zero-based idx)

* Turn eager rendering back on

* chore: Unused var

* Add storybook for observation with no location

* chore: Add types

* 🥳 Add icons to observations in report

* fix: patch hidden-mapbox so it does not hang

* fix: timeout PDF render

* fix: Cache category icon

* fix: Don't cache failed preview renders

* fix: pageIndex can end up with `null` values sometimes (not sure how!)

* fix: If preview is cancelled then bail

This can happen if the user changes the filter whilst the preview is
still loading. If the filter has changed, then we want to cancel async
functions that started with the previous filter

* fix: Catch infinite loop in preview

* feat: Use same map marker for observation as mapeo mobile

* fix: Fix patch to static map renderer

* fix: Only show fields with answers in report

* chore: Add Story for testing long preset names

* Add disabled state styling to page navigation buttons

* fix: Use new instance of ract-pdf for each render

Using the same renderer instance and updating the container was causing
race conditions, and does not seem to change performance much.

* chore: Use forked react-pdf/renderer

Changes babel preset-env to reduce polyfills and transpilation

* chore: Remove hidden-mapbox patch