chore: fix CI builds when run from a forked repo
Currently CI will break if run on a PR from a fork. This is because
the secrets are passed as empty strings, and electron-builder sees this
as "not null" so still tries to sign the installers.

This change ensures that the environment variables are not set when
the Github action is triggered from a forked repo, so that electron
builder just builds an un-signed installer.