update electron-builder to v23.6.0
Prior to v23.0.2, error would occur on macOS due to python 2.7 being
removed by OS update (e.g.
https://github.com/electron-userland/electron-builder/issues/6726).

The alternative approach is updating the workflow to install and use
Python 2. Probably safer and less intrusive to do that, so may consider
switching to that approach if this upgrade causes other issues.